Minor reorganization in tests
